+++ applications/debug/ccid_test/ccid_test_app.c
@@ -12,6 +12,7 @@
 #include "iso7816/iso7816_atr.h"
 #include "iso7816/iso7816_response.h"
 
+#include "ccid_usb.h"
 #include "ccid_test_app_commands.h"
 
 typedef enum {
@@ -45,9 +46,9 @@ typedef enum {
 static void ccid_test_submenu_callback(void* context, uint32_t index) {
     furi_assert(context);
     if(index == CcidTestSubmenuIndexInsertSmartcard) {
-        furi_hal_usb_ccid_insert_smartcard();
+        ccid_usb_insert_smartcard();
     } else if(index == CcidTestSubmenuIndexRemoveSmartcard) {
-        furi_hal_usb_ccid_remove_smartcard();
+        ccid_usb_remove_smartcard();
     }
 }
 
@@ -125,9 +126,9 @@ int32_t ccid_test_app(void* p) {
     FuriHalUsbInterface* usb_mode_prev = furi_hal_usb_get_config();
     furi_hal_usb_unlock();
 
-    furi_check(furi_hal_usb_set_config(&usb_ccid, &app->ccid_cfg) == true);
+    furi_check(furi_hal_usb_set_config(&ccid_usb_interface, &app->ccid_cfg) == true);
     iso7816_handler_set_usb_ccid_callbacks();
-    furi_hal_usb_ccid_insert_smartcard();
+    ccid_usb_insert_smartcard();
 
     view_dispatcher_run(app->view_dispatcher);