+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 * Apps: **NFC Magic** - Gen2 CUID/static-nonce detection, Gen1 4b/7b UID, length-aware wipe & write guard (by @mishamyte)
 * Apps: Build tag (**5jul2026p2**) - **Check out more Apps updates and fixes by following** [this link](https://github.com/xMasterX/all-the-plugins/commits/dev)
 ## Other changes
-* OFW: CCID: add interrupt support
+* OFW: CCID: move USB layer from firmware HAL into ccid_test app
 * OFW: Storage Python script: add retry on file copy to Flipper
 * Apps: Update FindMy app
 * Fix BLE sync, fix possible delay related issues